JabLibArchitectureTests by inheritance
