+++ jablib/src/test/java/org/jabref/architecture/JabLibArchitectureTests.java
@@ -1,0 +1,10 @@
+package org.jabref.architecture;
+
+import org.jabref.support.CommonArchitectureTest;
+
+import com.tngtech.archunit.core.importer.ImportOption;
+import com.tngtech.archunit.junit.AnalyzeClasses;
+
+@AnalyzeClasses(packages = "org.jabref", importOptions = ImportOption.DoNotIncludeTests.class)
+public class JabLibArchitectureTests extends CommonArchitectureTest {
+}
