+++ test-support/src/main/java/module-info.java
@@ -2,6 +2,8 @@ open module org.jabref.testsupport {
     requires org.junit.jupiter.api;
     requires org.mockito;
     requires org.jabref.jablib;
+    requires com.tngtech.archunit;
+    requires com.tngtech.archunit.junit5.api;
 
     exports org.jabref.support;
 }
